% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_fdi_to_alphanumeric.R
\name{convert_fdi_to_alphanumeric}
\alias{convert_fdi_to_alphanumeric}
\title{Convert FDI notation to alphanumeric notation (UR/UL/LL/LR)}
\usage{
convert_fdi_to_alphanumeric(fdi)
}
\arguments{
\item{fdi}{A numeric or character vector with FDI tooth codes.}
}
\value{
A character vector with alphanumeric notation. Invalid codes return NA.
}
\description{
Converts permanent and primary FDI notation into alphanumeric notation.
}
